♻️ [Refactoring]: #140 [FE] selectTaskOutcome を src/lib/ → src/domains/task-selection/ に移動
`src/lib/` はライブラリラッパ用途（API クライアント / 外部 SDK ラッパ等）
であり、Task ドメインの companion API として `selectTaskOutcome` を
ここに置くのは CLAUDE.md 規約と不整合だった。`src/domains/` 配下の
既存規約（kebab-case フォルダ + camelCase テストファイル, e.g.
task-hierarchy/__tests__/taskHierarchy.behavior.test.ts）に合わせて
`src/domains/task-selection/` に移動し、テストファイル名も
`taskSelection.behavior.test.ts` にリネームする。

Refs #140
